version 0.2.0 - redesign

+++ gfs-downsized/src/gfs_fc_download.py
@@ -6,7 +6,7 @@ analysis and forecast data in a trailing 30-day window in the AWS Open Data Regi
 Download GFS forecast data
 https://nomads.ncep.noaa.gov/pub/data/nccf/com/gfs/ via HTTP
 """
-
+import logging
 import pygrib
 import os
 import json
@@ -14,7 +14,7 @@ from numpy import array as np_array
 from multiprocessing import Queue
 from scipy.interpolate import RegularGridInterpolator
 # internal
-from gfs_fc_aux import data_file, config, defined_kwargs
+from gfs_fc_aux import DATA_FILE, CONFIG #, defined_kwargs
 
 
 def write_forecast(
@@ -27,11 +27,11 @@ def write_forecast(
     :param forecast:
     :return: None
     """
-    if not os.path.exists(data_file):
-        with open(data_file, "w") as create_empty:
+    if not os.path.exists(DATA_FILE):
+        with open(DATA_FILE, "w") as create_empty:
             json.dump({}, create_empty)
-            os.chmod(data_file, 0o666)  # docker owner is root, anyone can delete
-    with open(data_file, "r+") as jsonFile:
+            os.chmod(DATA_FILE, 0o666)  # docker owner is root, anyone can delete
+    with open(DATA_FILE, "r+") as jsonFile:
         data = json.load(jsonFile)
         data[datetimestr] = forecast
         jsonFile.seek(0)  # rewind
@@ -80,22 +80,28 @@ def create_grid(
 
 def extract(
         target: str,
-        q: Queue = None
+        q: Queue = None,
+        keep_target: bool = False
 ) -> tuple[bool, dict] | None:
     """
     extract grib2 file according to select parameter
     :param target: full path
     :param q: queue per fc hour for multiprocessing
+    :param keep_target: keep target, if True
     :return:
     """
     fs: list = list()
     result: dict = dict()
 
-    coords = np_array([config['geo_coordinates']['latitude'],
-                       (config['geo_coordinates']['longitude'] + 360) % 360])
+    coords = np_array([CONFIG['geo_coordinates']['latitude'],
+                       (CONFIG['geo_coordinates']['longitude'] + 360) % 360])
 
     # open grib file
     fsss = pygrib.open(target)
+    # for i in fsss:
+    #     print(i)
+    #     for j in i.keys():
+    #         print(j, getattr(i,j))
     fsss.seek(0)
     item = fsss.read(1)[0]
     # date of creation
@@ -113,23 +119,28 @@ def extract(
                        lats=lats[:, 0],
                        lons=lons[0, :])
 
-    for item in config['parameter']:
-        params = defined_kwargs(
-            shortName=item.get('shortName'),
-            typeOfLevel=item.get('typeOfLevel'),
-            level=item.get('level'),
-            stepType=item.get('stepType')
-            # ... more keys here if applicable
-        )
-        try:
-            fs.extend(fsss.select(**params))
-        except ValueError:
-            print("Filter parameter ", params, "not found. Skipping ...")
-
+    # ToDo shortNames are not equal in idx and grib2 files for GFS (NOAA). This
+    #  seems to be an issue of pygrib, as optimized for ECMWF. According to
+    #  ncep.pmb.dataflow@noaa.gov https://github.com/NOAA-MDL/grib2io might
+    #  be better suited. Hence, for the moment, we unload all
+    #  parameters available. Shit happenz.
+    # for item in CONFIG['parameter']:
+    #     params = defined_kwargs(
+    #         shortName=item.get('shortName'),
+    #         typeOfLevel=item.get('typeOfLevel'),
+    #         level=item.get('level'),
+    #         stepType=item.get('stepType')
+    #         # ... more keys here if applicable
+    #     )
+    #     try:
+    #         fs.extend(fsss.select(**params))
+    #     except ValueError:
+    #         print("Filter parameter ", params, "not found. Skipping ...")
+    fs.extend(fsss.select())
     fsss.close()
 
     for item in fs:
-        print(item["shortName"], item)
+        print(item["shortName"], "->", item)
         data, lats, lons = item.data(**grid)
         nearest_neighbor = RegularGridInterpolator(
             (lats[:, 0], lons[0, :]),
@@ -137,18 +148,35 @@ def extract(
             method='linear'
         )
         value_at_coordinates = list(nearest_neighbor(coords))[0]
+
+        # key is somewhat crummy
+        combine_dict_key = ("{}:{}:{}:{}"
+                            .format(item['name'],
+                                    item['stepType'],
+                                    item['typeOfLevel'],
+                                    item['level']))
         dt_str = "{}{:04d}".format(
             item['validityDate'],
             item['validityTime']
         )
-        result[item['name']] = {
+        result[combine_dict_key] = {
             "unit": item['units'],
             "time": [dt_str],
             "value": [value_at_coordinates]
         }
 
-    os.remove(target)  # ToDo keep or remove
-    print("Target file '{}' deleted".format(target))
+    # ToDo:
+    #  (Recitative) Thy hand, Belinda, darkness shades me,
+    #  On thy bosom let me rest,
+    #  More I would, but Death invades me;
+    #  Death is now a welcome guest.
+    #  (Aria) When I am laid, am laid in earth, May my wrongs create
+    #  No trouble, no trouble in thy breast;
+    #  Remember me, remember me, but ah! forget my fate.
+    #  Remember me, but ah! forget my fate.
+    if not keep_target:
+        os.remove(target)
+        logging.debug("Target file '{}' laid in earth".format(target))
 
     if q:
         q.put((date_creation_str, result))
